:+1: Error when `as const` is missing on isTupleOf
Now type error is shown when `as const` is required and missing on `isTupleOf` and `isReadonlyTupleOf`.
Note that it seems `predTup` that is directly assigned to `isTupleOf` or `isReadonlyTupleOf` are
automatically become `readonly` in recent TypeScript, meaning that `as const` is not required in that
case.

Additionally, empty `predTup` is no longer valid (type error will be shown) for `isTupleOf` and
`isReadonlyTupleOf`.