:boom: Rename `isRecord/isRecordOf/isRecordLike/isRecordLikeOf`
In unknownutil v3.16.0, we made slight adjustments to the behavior of
`isRecord/isRecordOf` to be more strict. However, upon realizing that users did
not expect such strictness in `isRecord/isRecordOf`, we introduced
`isRecordLike/isRecordLikeOf` as an alternative.

However, acknowledging that the naming convention might not be very intuitive,
we decided to roll back the changes to the behavior of `isRecord/isRecordOf` to
its state in unknownutil v3.15, abandoning the strictness, and introduced a new,
strictly enforced check called `isRecordObject/isRecordObjectOf`. As a result of
this modification, `isRecordLike/isRecordLikeOf` has been deprecated.

While these changes are disruptive, they represent a shift towards greater
flexibility and a return to the behavior of previous versions. Therefore, this
is considered a minor release.